Expose signal maps for popup menus
This allows to extend the signal types for custom base classes of any
of these types.

The interfaces are empty because I didn't care to figure out all signals
for the affected popup menu types, but still provided for forward
compatibility in order to allow signal types to be defined without
introducing new interfaces.