Fix types for PopupMenuBase.addMenuItem
It not only accepts PopupBaseMenuItem but also a few other types which
it explicitly checks for with instanceof and which do not inherit from
PopupBaseMenuItem and thus haven't type-checked so far.  We mirror the
checked types in an explicit enum.

Same goes for firstMenuItem which the implementation also narrows down
to two specific types via _getMenuItems()